Main

Description

Módulo para a realização da Tarefa 6 do projeto de LI1 em 2022/2023.

Documentation

main :: IO ()

desenha :: [Picture] -> GameState

A funcao desenha criar uma variavel do tipo GameState unindo uma lista de Pictures à variavel do GameState sem ela.

type GameState = ([Picture], (Integer, Float, Integer, [Integer], Jogo))

O estado do jogo composoto por : A lista de pictures que vão ser usadas na representação; Um inteiro que indica o menu atual (0 para menu inicial, 1 para menu de escolha de personagem, 3 para menu do jogo em si); Um Float que indica o personagem escolhido ou a ser escolhido (em fase de escolha: 0.0- nenhum personagem a ser selecionado, 0.1- pode ser selecionado o primeiro personagem, 0.2- pode ser selecionado o segundo personagem, 0.3- pode ser selecionado o terceiro personagem; personagem já escolhido: 1- primeiro personagem escolhido, 2- segundo personagem escolhido, 3- terceiro personagem escolhido); Um inteiro que dita se o jogador perdeu, colocou o jogo em pausa ou está vivo (0- está vivo, 1- perdeu, 3- pausa); Uma lista de inteiros que funciona como histórico de todos os pontos obtidos em todas as jogadas feitas; E uma variável do tipo Jogo;

estado_teste :: (Integer, Float, Integer, [Integer], Jogo)

jogo_inicial :: Jogo

displayMode :: Display

menu_choice :: GameState -> Picture

A funcao menu_choice vai indicar o menu a ser desenhado mediante os inteiros e floats recebidos e o Jogo. No caso de o primeiro inteiro ser 2 e o terceiro ser 0, não será desenhado um menu, mas sim o jogo em si (Mapa e Jogador).

start_game :: GameState -> Picture

A funcao start_game vai, a partir de um dado Game State com primeiro inteiro 2 e terceiro 0, desenhar o mapa de jogo, o jogador e o quadrado dos pontos.

pontos :: [Integer] -> Picture

A funcao pontos vai receber uma lista de pontos, que representa todos os pontos de cada tentativa feita. Esta funcao vai desenhar o quadrado onde os pontos serão expostos ao longo do jogo. É sempre representado o último elemento da lista pois este representa os pontos da tentativa mais atual.

menu_inicial :: Picture

A picture menu_inicial vai desenhar o menu inicial.

welcome :: Picture

A picture welcome vai desenhar o texto Welcome

start :: Picture

A picture start vai desenhar o texto "start"

menu_character :: [Picture] -> Float -> Picture

A funcao menu_character resulta em como vai ser desenhado o menu de escolha dos personagens, mediante um personagem estar a ser ou não selecionado.

char_chicken :: [Picture] -> Picture

A picture char_chicken vai desenhar a primeira escolha de personagem com um círculo azul à sua volta, ou seja, não está selecionado.

char_pig :: [Picture] -> Picture

A picture char_pig vai desenhar a segunda escolha de personagem com um círculo azul à sua volta, ou seja, não está selecionado.

char_cow :: [Picture] -> Picture

A picture char_cow vai desenhar a terceira escolha de personagem com um círculo azul à sua volta, ou seja, não está selecionado.

sel_char_chicken :: [Picture] -> Picture

A picture sel_char_chicken vai desenhar a primeira escolha de personagem com um círculo amarelo à sua volta, ou seja, está selecionado.

sel_char_pig :: [Picture] -> Picture

A picture sel_char_pig vai desenhar a segunda escolha de personagem com um círculo amarelo à sua volta, ou seja, está selecionado.

sel_char_cow :: [Picture] -> Picture

A picture sel_char_cow vai desenhar a terceira escolha de personagem com um círculo amarelo à sua volta, ou seja, está selecionado.

char_circle :: Picture

A picture char_circle desenha um circulo com contorno azul.

char_circleSelected :: Picture

A picture char_circleSelected desenha um circulo com contorno amarelo.

char_txtL :: Picture

A picture char_textL vai desenhar texto de explicação do menu de escolha de personagem.

char_txtS :: Picture

A picture char_textS vai desenhar texto de explicação do menu de escolha de personagem.

menu_pause :: Picture

A picture menu_pause vai desenhar o menu de pausa.

menu_lost :: [Integer] -> Picture

A picture menu_lost vai desenhar o menu em caso do jogo acabar. Vai receber também a lista de pontos e mediante a pontuacao da última jogada ser a melhor, um recorde, será desenhado um menu diferente do normal.

lost :: Picture

A picture lost vai desenhar o texto de aviso de fim do jogo.

restart :: Picture

A picture vai desenhar o texto com todas as mensagens que aparecem no menu de derrota.

restart1 :: Picture

A picture restart1 vai desenhar um texto de indicação.

restart2 :: Picture

A picture restart2 vai desenhar um texto de indicação.

restart3 :: Picture

A picture restart3 vai desenhar um texto de indicação.

new_record :: [Integer] -> Picture

A picture new_record vai desenhar o texto de aviso de novo recorde.

record :: [Integer] -> Picture

A picture record vai desenhar o texto a informar o maior recorde.

melhor_pontuacao :: [Integer] -> Integer

A funcao melhor_pontuacao vai receber a lista de todas as pontuções e retornar a maior delas.

drawCharacter :: [Picture] -> Float -> (Float, Float) -> Picture

A funcao drawCharacter vai desenhar um picture mediante um Float recebido. Essa picture é o personagem escolhido para jogar. Para além disso também recebe um par de Floats que representam a translação do personagem no mapa.

chickenP :: [Picture] -> Picture

A picture chickenP vai desenhar a primeira escolha de personagem que vai ser colocado no mapa, para se jogar.

pigP :: [Picture] -> Picture

A picture chickenP vai desenhar a segunda escolha de personagem que vai ser colocado no mapa, para se jogar.

cowP :: [Picture] -> Picture

A picture chickenP vai desenhar a terceira escolha de personagem que vai ser colocado no mapa, para se jogar.

picture_mapa :: [Picture] -> Mapa -> (Float, Float) -> Picture

A funcao picture_mapa vai receber a lista de pictures de linhas sem obstaculos e de linhas só com obstaculos juntas e desenhar uma única picture a partir dessa lista..

desenha_mapa :: [Picture] -> Mapa -> (Float, Float) -> [Picture]

A funcao desenha_mapa vai receber um mapa de modo desenhar a lista de pictures de linhas sem obstaculos e de linhas só com obstaculos juntas.

desenha_mapa_obstaculos :: [Picture] -> Mapa -> (Float, Float) -> [Picture]

A funcao desenha_mapa_obstaculos vai receber um mapa de modo a desenhar uma lista de pictures correspondente à lista de linhas do mapa só com obstaculos desenhados.

desenha_mapa_terrenos :: [Picture] -> Mapa -> (Float, Float) -> [Picture]

A funcao desenha_mapa_terrenos vai receber um mapa de modo a desenhar uma lista de pictures correspondente à lista de linhas do mapa, sem terrenos.

picture_linha_terrenos :: [Picture] -> Terreno -> (Float, Float) -> Int -> Picture

A funcao picture_linha_terrenos vai desenhar um só picture a partir da lista de pictures recebida.

desenha_linha_terrenos :: [Picture] -> Terreno -> (Float, Float) -> Int -> [Picture]

A funcao desenha_linha_terrenos desenha a lista de pictures do mesmmo terreno, ou seja uma linha. Esta funcao recebe também a largura do mapa para ter indicaçao de quantos quadrados são necessários desenhar e um par te Floats que indicam a translação feita a cada picture.

desenha_ter :: [Picture] -> Terreno -> (Float, Float) -> Picture

A funcao desenha_ter vai receber um terreno e um par de floats que vai associar a uma outra picture, mediante o terreno recebido. O par de Floats indica a translação feita à picture.

rioP :: [Picture] -> Picture

A picture rio P vai desenhar um quadrado do terreno rio.

estradaP :: [Picture] -> Picture

A picture estradaP vai desenhar um quadrado do terreno estrada.

relvaP :: [Picture] -> Picture

A picture relvaP vai desenhar um quadrado do terreno relva.

picture_linha_obstaculos :: [Picture] -> Velocidade -> [Obstaculo] -> (Float, Float) -> Picture

A funcao picture_linha_obstaculos vai desenhar um só picture a partir da lista de pictures recebida.

desenha_linha_obstaculos :: [Picture] -> Velocidade -> [Obstaculo] -> (Float, Float) -> [Picture]

A funcao desenha_linha_obstaculos desenha a lista de pictures da lista de terrenos recebida, ou seja uma linha. Esta funcao receber também a velocidade da linha e um par te Floats que indicam a translação feita a cada picture.

desenha_obs :: [Picture] -> Velocidade -> Obstaculo -> (Float, Float) -> Picture

A funcao desenha_obs vai receber um obstaculo e um par de floats que vai associar a uma outra picture, mediante o obstaculo recebido. O par de Floats indica a translação feita à picture. Recebe também a velocidade do terreno.

troncoP :: [Picture] -> Picture

A picture troncoP vai desenhar um quadrado de obstaculo Tronco.

carroP :: [Picture] -> Velocidade -> Picture

A picture carroP vai desenhar um quadrado de obstaculo Carro. Para além disso recebe a velocidade da linha, que vai ditar em que direção é desenhado o carro.

arvoreP :: [Picture] -> Picture

A picture arvoreP vai desenhar um quadrado de obstaculo Arvore.

nenhumP :: Picture

A picture nenhumP vai desenhar um espaço transparente, representando que não há obstaculo.

eventChange :: Event -> GameState -> GameState

A funcao eventChange vai alterar o GameState a partir de funções auxuliares relativas a cada caso de GameState recebido. Para além disso recebe o evento que vai ser usado.

menu_inicialChange :: Event -> GameState -> GameState

A funcao menu_inicialChange vai criar as mudanças de estado no menu inicial.

menu_characterChange :: Event -> GameState -> GameState

A funcao menu_characterChange vai determinar as mudanças de estado associadas ao menu de escolha de personagens.

wl_game :: Event -> GameState -> GameState

A funcao wl_game vai receber um estado representante do jogo a ser jogado e vai determinar se o jogo é valido ou se já está perdido. Caso jogo esteja perdido, leva para o menu de derrota. Para além disso também vai permitir o uso do menu de pausa.

playChange :: Jogada -> GameState -> GameState

A funcao playChange vai receber um jogada que vai ser aplicada no jogo do GameState associado a este.

contagem_pontos :: [Integer] -> Jogada -> [Integer]

A funcao contagem_pontos vai contar os pontos da tentiva que está a ser jogada, ou seja, soma um ponto caso o jogador passe para a linha seguinte ou subtrai um caso o jogador recue de linha. Caso a contagem esteja em zero, não podem ser retirados pontos.

associa_dir :: Event -> Jogada

A funcao associa_dir vai associar um evento a uma Jogada.

menu_lostChange :: Event -> GameState -> GameState

A funcao menu_lostChange vai determinar as mudanças associadas ao menu de derrota. Para além disso vai acresentar um novo elemento à lista de pontuações pois, se está no menu de derrota quer dizer que perdeu e portanto irá começar uma nova tentativa em que os pontos inicias têm de ser 0.

menu_pauseChange :: Event -> GameState -> GameState

A funcao menu_pauseChange vai determinar as mudanças no menu de pausa do jogo.

timeChange :: Float -> GameState -> GameState

A funcao timeChange vai determinar como a passagem do tempo afeta o GameState. Caso o jogador não esteja mesmo a jogar e esteja só nos menus, o tempo não afeta o estado. Caso esteja a jogar, o passar do tempo faz com que o estado se altere depois de se avaliar se o jogo do estado é valido para continuar.

timeChange_aux :: Float -> GameState -> GameState

A funcao timeChange_aux vai criar o efeito do mapa animado tal como do deslize do mapa do estado.