Tarefa2_2022li1g062

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/2023.

Documentation

proximosTerrenosValidos

Arguments

:: Mapa

-> [Terreno]

A função ’proximosTerrenosValidos’ calcula a lista de terrenos que poderao surgir na nova linha do mapa. Para esta funçao iremos ignorar os parametros relacionados com a velocidade do terreno Estrada e terreno Rio. Clicando em Mapa e em Terreno acima e possivel obter mais informaçoes relativamente a estas 2 funcoes.

Exemplos de utilização:

>>> proximosTerrenosValidos (Mapa 3 [(Estrada 3,[])])
[Relva,Rio 0,Estrada 0]

>>>  proximosTerrenosValidos (Mapa 3 [(Estrada 3,[Carro,Carro,Nenhum]), (Relva , [Arvore,Arvore,Nenhum])])
[Relva,Rio 0,Estrada 0]

>>> proximosTerrenosValidos (Mapa 2 [(Rio 7, [Tronco,Nenhum]), (Rio 1, [Tronco,Nenhum]),(Rio 2, [Tronco,Nenhum]),(Rio 3, [Tronco,Nenhum])])
[Estrada 0,Relva]

proximosObstaculosValidos :: Int -> Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

A função ’proximosObstaculosValidosauxiliar’ calcula os obstaculos que podem ser gerados para continuar uma dada linha do mapa. O valor inteiro corresponde a largura do mapa. Se o comprimento da lista de obstaculos atinge a largura do mapa entao mais nenhum obstaculo e possivel adicionar. Os obstaculos escolhidos devem ainda estar de acordo com o seu respetivo terreno. Clicando em Terreno e Obstaculo e possivel obter mais informaçoes relativamente a estas funcoes.

Exemplos de utilização:

>>> proximosObstaculosValidosauxiliar 10 (Estrada 3 , [Carro,Carro,Carro,Carro])
[Nenhum]

>>>  proximosObstaculosValidosauxiliar 3 (Estrada 3 , [Carro,Carro,Carro,Nenhum])
[Nenhum] 

>>>  proximosObstaculosValidosauxiliar 6 (Estrada 3 , [Carro,Carro,Carro,Carro])
[Nenhum]

>>> proximosObstaculosValidosauxiliar 6 (Estrada 3 , [])
[Nenhum,Carro]

>>>  proximosObstaculosValidosauxiliar 2 (Estrada 3 , [Carro,Carro,Carro,Carro])
[Nenhum]

proximosObstaculosValidosauxiliar :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

estendeMapa :: Mapa -> Int -> Mapa

A funcao ’estendeMapa’ tem como finalidade gerar e adicionar uma nova linha valida ao topo de um dado mapa. O valor inteiro deve estar entre [0,100] usado para acrescentar alguma pseudo-aleatoriedade a geracao a proxima nova linha. Acima encontram-se mais informaçoes sobre as funçoes Mapa.

Exemplos de utilização:

>>>  estendeMapa (Mapa 3 [(Relva , [])]) 3
Mapa 3 [(Relva,[]),(Relva,[Arvore,Arvore,Nenhum])]

Sobre...

Na funçao estendeMapa,